requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
groq==0.9.0
pymongo==4.8.0
python-dotenv==1.0.1
